requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.23.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
pytest>=7.0
open_clip_torch>=2.24.0
torch-fidelity>=0.3.0
diffusers>=0.25.0
scipy>=1.10.0